+++ RollbarNotifier/Sources/RollbarNotifier/include/RollbarLogger.h
@@ -4,6 +4,8 @@
 
 @class RollbarConfig;
 
+#import "RollbarLevel.h"
+
 @interface RollbarLogger : NSObject 
 
 /// Notifier's config object
@@ -32,12 +34,30 @@ NS_DESIGNATED_INITIALIZER;
 /// Captures a log entry
 /// @param level Rollbar error/log level
 /// @param message message
+/// @param data extra data
+/// @param context extra context
+- (void)log:(RollbarLevel)level
+    message:(NSString *)message
+       data:(NSDictionary<NSString *, id> *)data
+    context:(NSString *)context;
+
+/// Captures a log entry
+/// @param level Rollbar error/log level
 /// @param exception exception
 /// @param data extra data
 /// @param context extra context
-- (void)log:(NSString *)level
-    message:(NSString *)message
+- (void)log:(RollbarLevel)level
   exception:(NSException *)exception
+       data:(NSDictionary<NSString *, id> *)data
+    context:(NSString *)context;
+
+    /// Capture a log entry based on an NSError
+    /// @param level Rollbar error/log level
+    /// @param error an NSError
+    /// @param data extra data
+    /// @param context extra context
+- (void)log:(RollbarLevel)level
+      error:(NSError *)error
        data:(NSDictionary<NSString *, id> *)data
     context:(NSString *)context;